requirements.txt
# Core AI & Language Models
langchain>=0.3.0
langgraph>=0.0.21
langchain-ollama>=0.1.0
ollama>=0.1.8

# APIs & Web Requests
requests>=2.31.0
python-dotenv>=1.0.0
serpapi>=0.1.5

# Image Generation APIs
replicate>=0.22.0     # For Replicate API
openai>=1.0.0         # For OpenAI DALL-E

# Web App Framework
streamlit>=1.28.0
fastapi>=0.110.0
uvicorn>=0.29.0

# Social Media APIs
praw>=7.7.0           # For Reddit API
tweepy>=4.14.0        # For Twitter API

# Database & Memory
sqlite3               # Built-in Python module

# Optional (Debugging & Development)
rich>=13.7.0          # Pretty printing in terminal
